Fix incorrect warning message for HTTP status code
When a HTTP status code is a str, it is converted to bytes.

An assertion checks that the value is in fact bytes.

If the status code is neither bytes, nor a str, this assertion
will be triggered.

However, the error message incorrectly says "Status must be
string" instead.

This simple commit corrects the error message.